+++ carry/hang.go
@@ -24,7 +24,7 @@ var ProcessHang = func(market, symbol string) {
 		util.Notice(fmt.Sprintf(`%s %s [delay too long] %d`, market, symbol, delay))
 		return
 	}
-	priceDistance := 1 / math.Pow(10, float64(api.GetPriceDecimal(market, symbol)))
+	priceDistance := 1 / math.Pow(10, api.GetPriceDecimal(market, symbol))
 	checkDistance := priceDistance / 10
 	completeTick(market, symbol, tick, priceDistance, checkDistance)
 	setting := model.GetSetting(model.FunctionHang, market, symbol)
